Fix regression that prh was totally ignored now.
Commit 629e6cf brought in wrong use of file existence check:

	if (fs.exists (prhFile))

fs.exists() is supposed to take a callback to continue further operations.

And it turned out that fs.exists() is deprecated[*1] and should be
either fs.stat() or fs.statSync() for synchronized use like this line.

But instead of checking file existence, the entire validator instancing
can be simply made so that "if it fails to instantiate, just skip it".

Hence this simplification.

[*1] https://twitter.com/yosuke_furukawa/status/645260394592759808